backend/requirements.txt
# RobCrop Backend Requirements - CUDA 12.8 Compatible
# PyTorch will be installed via Docker with nightly cu128 builds

# FastAPI and web server
fastapi[all]
uvicorn[standard]==0.24.0

# Image processing and ML utilities
Pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# Additional ML utilities
scikit-learn==1.3.0
joblib==1.3.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Optional: Monitoring
prometheus-client==0.19.0

# Note: PyTorch is installed in Dockerfile with CUDA 12.8 support
# pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu128
